requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
supabase>=2.0.0
weaviate-client>=3.25.0
notion-client>=2.2.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
click>=8.1.0
httpx>=0.25.0
numpy>=1.24.0
pandas>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
types-requests>=2.31.0
click>=8.1.0
